requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
httpx==0.28.1
python-multipart==0.0.20
PyPDF2==3.0.1
python-docx==1.1.2
pandas==2.2.3
openpyxl==3.1.5
edge-tts==6.1.12
aiosqlite==0.20.0
